requirement.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Brotli==1.1.0
cffi==1.16.0
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.5
django-import-export==3.3.6
django-simple-history==3.6.0
django-weasyprint==2.3.0
et-xmlfile==1.1.0
fonttools==4.53.0
html5lib==1.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.3
pillow==10.3.0
pycparser==2.22
pydyf==0.10.0
pyphen==0.15.0
PyYAML==6.0.1
six==1.16.0
sqlparse==0.5.0
tablib==3.5.0
tinycss2==1.3.0
typing_extensions==4.12.0
tzdata==2024.1
weasyprint==61.2
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
zopfli==0.2.3
